Widen cc3d labels before offsetting the negative tail
cc3d returns uint16 labels when the count fits, and NumPy 2 keeps the
narrow dtype through 'lab + offset'. Once a bi-sided map had more than
65535 clusters in total, negative-tail labels wrapped onto positive ones,
so sizes and label map disagreed. Found by clusterizing a loose threshold
in the viewer, where it surfaced as an IndexError while renumbering.

+++ fastfuncstuff/stats/cluster.py
@@ -144,7 +144,10 @@ def _cluster_extent_mass_one(
             all_masses.append(masses)
             # Splice into combined label map with offset
             mask_nonzero = lab > 0
-            combined_labels[mask_nonzero] = lab[mask_nonzero] + offset
+            # Widened before the offset: cc3d returns uint16 when it can, and
+            # NumPy 2 keeps the narrow type, so past 65535 total the negative
+            # tail's labels wrapped onto the positive tail's.
+            combined_labels[mask_nonzero] = lab[mask_nonzero].astype(np.int32) + offset
             offset += n_lab
         if not all_sizes:
             empty = np.zeros(0, dtype=np.int64)
